requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-tavily>=0.1.0
langchain-mcp-adapters>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
